Extract PlayControls from Game.tsx.
